Feat: Add blank pages for missing links
- Installed `react-router-dom` to handle routing.
- Created blank page components for Platform, Consulting, Case Studies, Insights, and Contact.
- Set up routing in `App.tsx` to render the correct page based on the URL.
- Moved the existing content of `App.tsx` into a new `Home.tsx` component.
- Updated the `Header` component to use `Link` from `react-router-dom`.
