+++ components/apiKey.ts
@@ -1,0 +1,1 @@
+export const apiKey = "AIzaSyCIZVF-0wwtlhw75cl2H7a0vLFC3qCdP4o";
