+++ pages/CaseStudies.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CaseStudies = (): React.ReactNode => {
+  return (
+    <div className="text-white">
+      <h1>Case Studies Page</h1>
+    </div>
+  );
+};
+
+export default CaseStudies;
