+++ pages/Consulting.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Consulting = (): React.ReactNode => {
+  return (
+    <div className="text-white">
+      <h1>Consulting Page</h1>
+    </div>
+  );
+};
+
+export default Consulting;
