+++ pages/Contact.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Contact = (): React.ReactNode => {
+  return (
+    <div className="text-white">
+      <h1>Contact Page</h1>
+    </div>
+  );
+};
+
+export default Contact;
