+++ pages/Insights.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Insights = (): React.ReactNode => {
+  return (
+    <div className="text-white">
+      <h1>Insights Page</h1>
+    </div>
+  );
+};
+
+export default Insights;
