+++ pages/Platform.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Platform = (): React.ReactNode => {
+  return (
+    <div className="text-white">
+      <h1>Platform Page</h1>
+    </div>
+  );
+};
+
+export default Platform;
